Got Units to consist between battles, but now attacks only work once

+++ Assets/Scripts/Battle/BattleHUD.cs
@@ -17,7 +17,7 @@ public class BattleHUD : MonoBehaviour
     public Button magicButton;
     public Button backButton;
 
-    private int speedMod;
+    private float speedMod;
 
 
     private void Start()
